Made use of async and await on server side

+++ server/server.js
@@ -65,24 +65,22 @@ app.get('/todos/:id', authenticate, (req, res) => {
 });
 
 // DELETE a todo by a specific id
-app.delete('/todos/:id', authenticate, (req, res) => {
+app.delete('/todos/:id', authenticate, async (req, res) => {
   var id = req.params.id;
 
   if(!ObjectID.isValid(id)){
     return res.status(404).send();
   }
 
-  Todo.findOneAndRemove({
-    _id: id,
-    _creator: req.user._id
-  }).then((todo) => {
-    if(!todo){
+  try {
+    const todo = await Todo.findOneAndRemove({_id: id, _creator: req.user._id});
+    if(!todo) {
       return res.status(404).send();
     }
     res.send({todo});
-  }).catch((e) => {
-    res.status(400).send(e);
-  });
+  } catch(e) {
+    res.status(404).send();
+  }
 });
 
 
@@ -114,42 +112,40 @@ app.patch('/todos/:id', authenticate, (req, res) => {
 
 
 // POST /users
-app.post("/users", (req, res) => {
-  var body = _.pick(req.body, ["email", "password"]);
-  var user = new User(body);
-
-  user.save().then(() => {
-    return user.generateAuthToken();
-    // this user is not the same as the user sent in the promise below
-    // res.send(user);
-  }).then((token) => {
-    // we have to add on the header. we have to send the token back as an http response header
+app.post("/users", async (req, res) => {
+  try {
+    var body = _.pick(req.body, ["email", "password"]);
+    var user = new User(body);
+    await user.save();
+    var token = user.generateAuthToken();
     res.header('x-auth', token).send(user);
-  }).catch(e => res.status(400).send(e));
+  } catch (e) {
+    res.status(400).send(e);
+  }
 });
 
 app.get('/users/me', authenticate, (req, res, next) => {
   res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-  var body = _.pick(req.body, ['email', 'password']);
-
-  User.findByCredentials(body.email, body.password).then((user) => {
-    return user.generateAuthToken().then((token) => {
-      res.header('x-auth', token).send(user);
-    });
-  }).catch((e) => {
+app.post('/users/login', async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password']);
+    const user = await User.findByCredentials(body.email, body.password);
+    const token = await user.generateAuthToken();
+    res.header('x-auth', token).send(user);
+  } catch (e) {
     res.status(400).send();
-  });
+  }
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-  req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+  try {
+    await req.user.removeToken(req.token);
     res.status(200).send();
-  }, () => {
+  } catch (e) {
     res.status(400).send();
-  })
+  }
 });
